rainshop/requirements.txt
amqp==5.0.6
asgiref==3.4.1
astroid==2.7.2
atomicwrites==1.4.0
attrs==21.2.0
backports.entry-points-selectable==1.1.0
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.2
dj-database-url==0.5.0
dj-rest-auth==2.1.11
Django==3.2.6
gunicorn==20.1.0
django-autoslug==1.9.8
django-cities-light==3.8.1
django-cors-headers==3.7.0
django-countries==7.2.1
django-db-logger==0.1.10
django-debug-toolbar==3.2
django-environ==0.4.5
django-extensions==3.1.3
django-filter==2.4.0
django-phone-field==1.8.1
django-redis==5.0.0
django-rest-swagger==2.2.0
django-settings-json==1.0.2
django-simple-history==3.0.0
django-timezone-field==4.2.1
django-watchman==1.2.0
djangorestframework==3.12.4
djangorestframework-api-key==2.0.0
djangorestframework-simplejwt==4.7.2
drf-yasg==1.20.0
et-xmlfile==1.1.0
factory-boy==3.2.0
Faker==8.11.0
filelock==3.0.12
idna==3.2
inflection==0.5.1
iniconfig==1.1.1
isort==5.9.3
itypes==1.2.0
Jinja2==3.0.1
kombu==5.1.0
lazy-object-proxy==1.6.0
MarkupPy==1.14
MarkupSafe==2.0.1
mccabe==0.6.1
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.7
packaging==21.0
Pillow==8.3.1
platformdirs==2.2.0
pluggy==0.13.1
progressbar2==3.53.1
prompt-toolkit==3.0.19
psycopg2==2.9.1
py==1.10.0
PyJWT==2.1.0
pylint==2.10.2
pylint-django==2.4.4
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==6.2.4
pytest-cov==2.12.1
pytest-django==4.4.0
python-dateutil==2.8.2
python-utils==2.5.6
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.26.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
simplejson==3.17.3
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
tox==3.24.1
Unidecode==1.2.0
uritemplate==3.0.1
urllib3==1.26.6
uuid==1.30
vine==5.0.0
virtualenv==20.7.1
wcwidth==0.2.5
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
